+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Added support for custom Message-ID headers in Mailer in Mailgun implementation. Mailgun reused same Message-ID for all emails within one batch which could cause unexpected behavior. remp/remp#801
 
+### [Sso]
+
+- Models from `\App` namespace moved to `App\Models` (compatibility with current Laravel conventions).
+
 ## [0.18.0] - 2021-01-14
 
 ### [Beam]